Replace custom traits by XTypes' TypeSupport  (#347)
* add trial file

* fix padding

* use md5_collection

* use recursion

* use Write trait

* add Large key test

* add result

* Add key field to type support derive

* use xtypes(key)

* add TypeSupport bound

* make python bindings compile

* add get_instance_handle_from_serialized_foo

* impl TypeSupport for built-ins

* fix one test

* revert data_writer_api

* add get_instance_handle_from_serialized_key

* Delete DdsKey DdsHasKey and DdsTypeXml

* add key to discovery tests

* fix key_field of builtins

* add type_obejct match arms

* fix tests

* generalize deserializer

* Add id attribute parsing to type support

* add MutableStruct test

* fix type for builtins

* Allow automatic derive of mutable struct with field id

* Revert test changes

* Use TypeIdentifier instead of DynamicType

* Fix failing test

* use own deserializer each pid field

* unify matches

* simplyfy go_to_pid_le

* use TypeIdentifier

* add iterator

* adjust sequence types

* Fix typo

* fix tests and remove warnings

* add error handling to iterator

* add basic types test stub

* move new stuff

* move new stuff

* fix clippy

* add basic types

* remove test foo_with_specialized_type_support_should_read_and_write

* fix padding of serialized key

* remove obsolete test

* Remove configuration on test

* Add type support for python bindings

* Use tryfrom to create dynamic type representation

* Fix python tests

* Fix clippy

* Store Python type in global storage

---------

+++ dds/src/xtypes/deserializer.rs
@@ -15,12 +15,12 @@ pub trait DeserializeAppendableStruct<'a> {
 pub trait DeserializeMutableStruct<'a> {
     fn deserialize_field<T: XTypesDeserialize<'a>>(
         &mut self,
-        pid: u16,
+        pid: u32,
         name: &str,
     ) -> Result<T, XTypesError>;
     fn deserialize_optional_field<T: XTypesDeserialize<'a>>(
         &mut self,
-        pid: u16,
+        pid: u32,
         name: &str,
     ) -> Result<Option<T>, XTypesError>;
 }
